Refactor: Split API modules into public/internal (#1083)
* split gundeck API

* split cannon API

* split galley API

* move cargohold API to module Public

* split brig API

* split proxy API

* formatting for stern
