+++ tools/stern/src/Stern/API.hs
@@ -81,11 +81,16 @@ start o = do
 
 sitemap :: Routes Doc.ApiBuilder Handler ()
 sitemap = do
-  ---------------------------------------------------------------------------
+  routes
+  apiDocs
+
+routes :: Routes Doc.ApiBuilder Handler ()
+routes = do
   -- Begin Internal
 
   get "/i/status" (continue $ const $ return empty) true
   head "/i/status" (continue $ const $ return empty) true
+
   -- End Internal
 
   post "/users/:uid/suspend" (continue suspendUser) $
@@ -96,6 +101,7 @@ sitemap = do
       Doc.description "User ID"
     Doc.response 200 "User successfully suspended" Doc.end
     Doc.response 400 "Bad request" (Doc.model Doc.errorModel)
+
   post "/users/:uid/unsuspend" (continue unsuspendUser) $
     capture "uid"
   document "POST" "users/:uid/unsuspend" $ do
@@ -104,6 +110,7 @@ sitemap = do
       Doc.description "User ID"
     Doc.response 200 "User successfully unsuspended" Doc.end
     Doc.response 400 "Bad request" (Doc.model Doc.errorModel)
+
   get "/users" (continue usersByEmail) $
     param "email"
   document "GET" "users" $ do
@@ -111,6 +118,7 @@ sitemap = do
     Doc.parameter Doc.Query "email" Doc.string' $
       Doc.description "Email address"
     Doc.response 200 "List of users" Doc.end
+
   get
     "/users"
     (continue usersByPhone)
@@ -120,6 +128,7 @@ sitemap = do
     Doc.parameter Doc.Query "phone" Doc.string' $
       Doc.description "Phone number"
     Doc.response 200 "List of users" Doc.end
+
   get "/users" (continue usersByIds) $
     param "ids"
   document "GET" "users" $ do
@@ -127,6 +136,7 @@ sitemap = do
     Doc.parameter Doc.Query "ids" Doc.string' $
       Doc.description "ID of the user"
     Doc.response 200 "List of users" Doc.end
+
   get "/users" (continue usersByHandles) $
     param "handles"
   document "GET" "users" $ do
@@ -134,6 +144,7 @@ sitemap = do
     Doc.parameter Doc.Query "handles" Doc.string' $
       Doc.description "Handle of the user"
     Doc.response 200 "List of users" Doc.end
+
   get "/users/:uid/connections" (continue userConnections) $
     capture "uid"
   document "GET" "users/:uid/connections" $ do
@@ -141,6 +152,7 @@ sitemap = do
     Doc.parameter Doc.Path "uid" Doc.bytes' $
       description "User ID"
     Doc.response 200 "List of user's connections" Doc.end
+
   get "/users/connections" (continue usersConnections) $
     param "ids"
   document "GET" "users/connections" $ do
@@ -148,6 +160,7 @@ sitemap = do
     Doc.parameter Doc.Query "ids" Doc.string' $
       Doc.description "IDs of the users"
     Doc.response 200 "List of users connections" Doc.end
+
   get "/users/:uid/search" (continue searchOnBehalf) $
     capture "uid"
       .&. def "" (query "q")
@@ -163,6 +176,7 @@ sitemap = do
       description "Number of results to return"
       optional
     Doc.response 200 "List of users" Doc.end
+
   post "/users/revoke-identity" (continue revokeIdentity) $
     param "email" ||| phoneParam
   document "POST" "revokeIdentity" $ do
@@ -181,6 +195,7 @@ sitemap = do
       Doc.optional
     Doc.response 200 "Identity revoked or not verified / taken." Doc.end
     Doc.response 400 "Bad request" (Doc.model Doc.errorModel)
+
   put "/users/:uid/email" (continue changeEmail) $
     contentType "application" "json"
       .&. capture "uid"
@@ -196,6 +211,7 @@ sitemap = do
       Doc.description "JSON body"
     Doc.response 200 "Change of email address initiated." Doc.end
     Doc.response 400 "Bad request" (Doc.model Doc.errorModel)
+
   put "/users/:uid/phone" (continue changePhone) $
     contentType "application" "json"
       .&. capture "uid"
@@ -211,6 +227,7 @@ sitemap = do
       Doc.description "JSON body"
     Doc.response 200 "Change of phone number initiated." Doc.end
     Doc.response 400 "Bad request" (Doc.model Doc.errorModel)
+
   delete "/users/:uid" (continue deleteUser) $
     capture "uid"
       .&. (query "email" ||| phoneParam)
@@ -227,6 +244,7 @@ sitemap = do
       Doc.optional
     Doc.response 200 "Account deleted" Doc.end
     Doc.response 400 "Bad request" (Doc.model Doc.errorModel)
+
   delete "/teams/:tid" (continue deleteTeam) $
     capture "tid"
       .&. query "email"
@@ -255,6 +273,7 @@ sitemap = do
       Doc.optional
     Doc.response 200 "The email/phone IS blacklisted" Doc.end
     Doc.response 404 "The email/phone is NOT blacklisted" Doc.end
+
   post "/users/blacklist" (continue addBlacklist) $
     (query "email" ||| phoneParam)
   document "POST" "addToBlacklist" $ do
@@ -266,6 +285,7 @@ sitemap = do
       Doc.description "A phone to add"
       Doc.optional
     Doc.response 200 "Operation succeeded" Doc.end
+
   delete "/users/blacklist" (continue deleteFromBlacklist) $
     (query "email" ||| phoneParam)
   document "DELETE" "deleteFromBlacklist" $ do
@@ -277,6 +297,7 @@ sitemap = do
       Doc.description "A phone to remove"
       Doc.optional
     Doc.response 200 "Operation succeeded" Doc.end
+
   get "/teams" (continue getTeamInfoByMemberEmail) $
     param "email"
   document "GET" "getTeamInfoByMemberEmail" $ do
@@ -284,6 +305,7 @@ sitemap = do
     Doc.parameter Doc.Query "email" Doc.string' $
       Doc.description "A verified email address"
     Doc.response 200 "Team Information" Doc.end
+
   get "/teams/:tid" (continue getTeamInfo) $
     capture "tid"
   document "GET" "getTeamInfo" $ do
@@ -291,6 +313,7 @@ sitemap = do
     Doc.parameter Doc.Path "tid" Doc.bytes' $
       description "Team ID"
     Doc.response 200 "Team Information" Doc.end
+
   get "/teams/:tid/admins" (continue getTeamAdminInfo) $
     capture "tid"
   document "GET" "getTeamAdminInfo" $ do
@@ -298,6 +321,7 @@ sitemap = do
     Doc.parameter Doc.Path "tid" Doc.bytes' $
       description "Team ID"
     Doc.response 200 "Team Information about Owners and Admins" Doc.end
+
   -- feature flags
 
   get "/teams/:tid/features/legalhold" (continue (liftM json . Intra.getLegalholdStatus)) $
@@ -309,6 +333,7 @@ sitemap = do
     Doc.returns Doc.docSetLegalHoldStatus
     Doc.response 200 "Legalhold status" Doc.end
     Doc.returns Doc.bool'
+
   put "/teams/:tid/features/legalhold" (continue setLegalholdStatus) $
     contentType "application" "json"
       .&. capture "tid"
@@ -320,6 +345,7 @@ sitemap = do
     Doc.body Doc.docSetLegalHoldStatus $
       Doc.description "JSON body"
     Doc.response 200 "Legalhold status" Doc.end
+
   get "/teams/:tid/features/sso" (continue (liftM json . Intra.getSSOStatus)) $
     capture "tid"
   document "GET" "getSSOStatus" $ do
@@ -328,6 +354,7 @@ sitemap = do
       description "Team ID"
     Doc.returns Doc.docSetSSOStatus
     Doc.response 200 "SSO status" Doc.end
+
   put "/teams/:tid/features/sso" (continue setSSOStatus) $
     contentType "application" "json"
       .&. capture "tid"
@@ -339,16 +366,8 @@ sitemap = do
     Doc.body Doc.docSetSSOStatus $
       Doc.description "JSON body"
     Doc.response 200 "SSO status" Doc.end
-  --- Swagger ---
-  get
-    "/stern/api-docs"
-    ( \(_ ::: url) k ->
-        let doc = mkSwaggerApi (decodeLatin1 url) Doc.sternModels sitemap
-         in k $ json doc
-    )
-    $ accept "application" "json"
-      .&. query "base_url"
-  -- The following endpoint are only relevant internally at Wire
+
+  -- The following endpoints are only relevant internally at Wire
 
   get "/teams/:tid/invoices/:inr" (continue getTeamInvoice) $
     capture "tid"
@@ -362,6 +381,7 @@ sitemap = do
     Doc.parameter Doc.Path "inr" Doc.string' $
       Doc.description "Invoice Number"
     Doc.response 307 "Redirect to PDF download" Doc.end
+
   get "/teams/:tid/billing" (continue getTeamBillingInfo) $
     capture "tid"
   document "GET" "getTeamBillingInfo" $ do
@@ -372,6 +392,7 @@ sitemap = do
     Doc.response 200 "Team Billing Information" Doc.end
     Doc.response 404 "No team or no billing info for given team" Doc.end
     Doc.returns (Doc.ref Doc.teamBillingInfo)
+
   put "/teams/:tid/billing" (continue updateTeamBillingInfo) $
     contentType "application" "json"
       .&. capture "tid"
@@ -387,6 +408,7 @@ sitemap = do
       Doc.description "JSON body"
     Doc.response 200 "Updated Team Billing Information" Doc.end
     Doc.returns (Doc.ref Doc.teamBillingInfo)
+
   post "/teams/:tid/billing" (continue setTeamBillingInfo) $
     contentType "application" "json"
       .&. capture "tid"
@@ -404,6 +426,7 @@ sitemap = do
       Doc.description "JSON body"
     Doc.response 200 "Updated Team Billing Information" Doc.end
     Doc.returns (Doc.ref Doc.teamBillingInfo)
+
   get "/i/consent" (continue getConsentLog) $
     param "email"
   document "GET" "getConsentLog" $ do
@@ -413,6 +436,7 @@ sitemap = do
       Doc.description "An email address"
     Doc.response 200 "Consent Log" Doc.end
     Doc.response 403 "Access denied! There is a user with this email address" Doc.end
+
   get "/i/user/meta-info" (continue getUserData) $
     param "id"
   document "GET" "getUserMetaInfo" $ do
@@ -421,6 +445,17 @@ sitemap = do
     Doc.parameter Doc.Query "id" Doc.bytes' $ do
       Doc.description "A user's ID"
     Doc.response 200 "Meta Info" Doc.end
+
+apiDocs :: Routes a Handler ()
+apiDocs = do
+  get
+    "/stern/api-docs"
+    ( \(_ ::: url) k ->
+        let doc = mkSwaggerApi (decodeLatin1 url) Doc.sternModels routes
+         in k $ json doc
+    )
+    $ accept "application" "json"
+      .&. query "base_url"
 
 -----------------------------------------------------------------------------
 -- Handlers
